% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{AvantGardeDIA_GlobalRefinement}
\alias{AvantGardeDIA_GlobalRefinement}
\title{AvantGardeDIA_GlobalRefinement}
\usage{
AvantGardeDIA_GlobalRefinement(index, D)
}
\description{
This function is AvantGardeDIA_GlobalRefinement
}
\examples{
AvantGardeDIA_GlobalRefinement()
}
\keyword{AvantGardeDIA}
