% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{AvantGardeDIA_InChunks_DB}
\alias{AvantGardeDIA_InChunks_DB}
\title{AvantGardeDIA_InChunks_DB}
\usage{
AvantGardeDIA_InChunks_DB(D.file.name, RefinementWorkflow)
}
\description{
This function is AvantGardeDIA_InChunks_DB
}
\examples{
AvantGardeDIA_InChunks_DB()
}
\keyword{AvantGardeDIA}
