% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{AvantGardeDIA_PeakBoundariesAndReScore}
\alias{AvantGardeDIA_PeakBoundariesAndReScore}
\title{AvantGardeDIA_PeakBoundariesAndReScore}
\usage{
AvantGardeDIA_PeakBoundariesAndReScore(index, D)
}
\description{
This function is AvantGardeDIA_PeakBoundariesAndReScore
}
\examples{
AvantGardeDIA_PeakBoundariesAndReScore()
}
\keyword{AvantGardeDIA}
