% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{AvantGardeDIA_ReScore}
\alias{AvantGardeDIA_ReScore}
\title{AvantGardeDIA_ReScore}
\usage{
AvantGardeDIA_ReScore(index, D)
}
\description{
This function is AvantGardeDIA_ReScore
}
\examples{
AvantGardeDIA_ReScore()
}
\keyword{AvantGardeDIA}
