% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{AvantGardeDIA_SkorPlots}
\alias{AvantGardeDIA_SkorPlots}
\title{AvantGardeDIA_SkorPlots}
\usage{
AvantGardeDIA_SkorPlots(index, D)
}
\description{
This function is AvantGardeDIA_SkorPlots
}
\examples{
AvantGardeDIA_SkorPlots()
}
\keyword{AvantGardeDIA}
