% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{AvantGardeDIA_TransitionRefinementAndReScore}
\alias{AvantGardeDIA_TransitionRefinementAndReScore}
\title{AvantGardeDIA_TransitionRefinementAndReScore}
\usage{
AvantGardeDIA_TransitionRefinementAndReScore(index, D)
}
\description{
This function is AvantGardeDIA_TransitionRefinementAndReScore
}
\examples{
AvantGardeDIA_TransitionRefinementAndReScore()
}
\keyword{AvantGardeDIA}
