% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{AvG_writeParamsUsed}
\alias{AvG_writeParamsUsed}
\title{AvG_writeParamsUsed}
\usage{
AvG_writeParamsUsed(SubSetOfParams)
}
\description{
This function is AvG_writeParamsUsed
}
\examples{
AvG_writeParamsUsed()
}
\keyword{AvantGardeDIA}
