% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Chromatographic.DotP}
\alias{Chromatographic.DotP}
\title{Chromatographic.DotP}
\usage{
Chromatographic.DotP(list, Chrom.IsPotentialPeak)
}
\description{
This function is Chromatographic.DotP
}
\examples{
Chromatographic.DotP()
}
\keyword{AvantGardeDIA}
