% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Chromatographic.IntProductScore}
\alias{Chromatographic.IntProductScore}
\title{Chromatographic.IntProductScore}
\usage{
Chromatographic.IntProductScore(list)
}
\description{
This function is Chromatographic.IntProductScore
}
\examples{
Chromatographic.IntProductScore()
}
\keyword{AvantGardeDIA}
