% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Chromatographic.IntScore}
\alias{Chromatographic.IntScore}
\title{Chromatographic.IntScore}
\usage{
Chromatographic.IntScore(list)
}
\description{
This function is Chromatographic.IntScore
}
\examples{
Chromatographic.IntScore()
}
\keyword{AvantGardeDIA}
