% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Chromatographic.IsPotentialPeak}
\alias{Chromatographic.IsPotentialPeak}
\title{Chromatographic.IsPotentialPeak}
\usage{
Chromatographic.IsPotentialPeak(list, MinimalNumberOfConsecutivePoints)
}
\description{
This function is Chromatographic.IsPotentialPeak
}
\examples{
Chromatographic.IsPotentialPeak()
}
\keyword{AvantGardeDIA}
