% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Chromatographic.MPRA}
\alias{Chromatographic.MPRA}
\title{Chromatographic.MPRA}
\usage{
Chromatographic.MPRA(list, Chrom.IsPotentialPeak)
}
\description{
This function is Chromatographic.MPRA
}
\examples{
Chromatographic.MPRA()
}
\keyword{AvantGardeDIA}
