% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Data.Loader_DB}
\alias{Data.Loader_DB}
\title{Data.Loader_DB}
\usage{
Data.Loader_DB(index, D)
}
\description{
This function is Data.Loader_DB
}
\examples{
Data.Loader_DB()
}
\keyword{AvantGardeDIA}
