% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{FindTranswithConsecutivePoints}
\alias{FindTranswithConsecutivePoints}
\title{FindTranswithConsecutivePoints}
\usage{
FindTranswithConsecutivePoints(Norm.Chrom,
  MinimalIntensityPercentagePerTransition, MinimalNumberOfConsecutivePoints)
}
\description{
This function is FindTranswithConsecutivePoints
}
\examples{
FindTranswithConsecutivePoints()
}
\keyword{AvantGardeDIA}
