% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{GA.Fitness}
\alias{GA.Fitness}
\title{GA.Fitness}
\usage{
GA.Fitness(y, Norm.Chrom, Transition.Area, Transition.Rank, MassErrors)
}
\description{
This function is GA.Fitness
}
\examples{
GA.Fitness()
}
\keyword{AvantGardeDIA}
