% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Indexing}
\alias{Indexing}
\title{Indexing}
\usage{
Indexing(D)
}
\description{
This function is Indexing
}
\examples{
Indexing()
}
\keyword{AvantGardeDIA}
