% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Intensity.Fitness}
\alias{Intensity.Fitness}
\title{Intensity.Fitness}
\usage{
Intensity.Fitness(Transition.Area, y)
}
\description{
This function is Intensity.Fitness
}
\examples{
Intensity.Fitness()
}
\keyword{AvantGardeDIA}
