% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Intensity.Fitness.Report}
\alias{Intensity.Fitness.Report}
\title{Intensity.Fitness.Report}
\usage{
Intensity.Fitness.Report(Transition.Area, y)
}
\description{
This function is Intensity.Fitness.Report
}
\examples{
Intensity.Fitness.Report()
}
\keyword{AvantGardeDIA}
