% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{LaunchParallelTasks_ReadFromDB}
\alias{LaunchParallelTasks_ReadFromDB}
\title{LaunchParallelTasks_ReadFromDB}
\usage{
LaunchParallelTasks_ReadFromDB(ParamsFile, RefinementWorkflow)
}
\description{
This function is LaunchParallelTasks_ReadFromDB
}
\examples{
LaunchParallelTasks_ReadFromDB()
}
\keyword{AvantGardeDIA}
