% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Library.dotp}
\alias{Library.dotp}
\title{Library.dotp}
\usage{
Library.dotp(list)
}
\description{
This function is Library.dotp
}
\examples{
Library.dotp()
}
\keyword{AvantGardeDIA}
