% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{LinearEquationByRange}
\alias{LinearEquationByRange}
\title{LinearEquationByRange}
\usage{
LinearEquationByRange(x, x_range, y_range)
}
\description{
This function is LinearEquationByRange
}
\examples{
LinearEquationByRange()
}
\keyword{AvantGardeDIA}
