% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{MassError.Score.Fitness_informsMPRA}
\alias{MassError.Score.Fitness_informsMPRA}
\title{MassError.Score.Fitness_informsMPRA}
\usage{
MassError.Score.Fitness_informsMPRA(MassErrors, y)
}
\description{
This function is MassError.Score.Fitness_informsMPRA
}
\examples{
MassError.Score.Fitness_informsMPRA()
}
\keyword{AvantGardeDIA}
