% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{MetaDataLibraryConcatenator}
\alias{MetaDataLibraryConcatenator}
\title{MetaDataLibraryConcatenator}
\usage{
MetaDataLibraryConcatenator(MetaData1, MetaData2)
}
\description{
This function is MetaDataLibraryConcatenator
}
\examples{
MetaDataLibraryConcatenator()
}
\keyword{AvantGardeDIA}
