% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{MPRA.Score.dotp}
\alias{MPRA.Score.dotp}
\title{MPRA.Score.dotp}
\usage{
MPRA.Score.dotp(list)
}
\description{
This function is MPRA.Score.dotp
}
\examples{
MPRA.Score.dotp()
}
\keyword{AvantGardeDIA}
