% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{MPRA_y_SpectLib.Fitness}
\alias{MPRA_y_SpectLib.Fitness}
\title{MPRA_y_SpectLib.Fitness}
\usage{
MPRA_y_SpectLib.Fitness(Transition.Area, y)
}
\description{
This function is MPRA_y_SpectLib.Fitness
}
\examples{
MPRA_y_SpectLib.Fitness()
}
\keyword{AvantGardeDIA}
