% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{MPRA_y_SpectLib.Fitness_informed}
\alias{MPRA_y_SpectLib.Fitness_informed}
\title{MPRA_y_SpectLib.Fitness_informed}
\usage{
MPRA_y_SpectLib.Fitness_informed(Transition.Area, y, Best_reps,
  Best_reps_MassErrors)
}
\description{
This function is MPRA_y_SpectLib.Fitness_informed
}
\examples{
MPRA_y_SpectLib.Fitness_informed()
}
\keyword{AvantGardeDIA}
