% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{MPRA_y_SpectLib.Report}
\alias{MPRA_y_SpectLib.Report}
\title{MPRA_y_SpectLib.Report}
\usage{
MPRA_y_SpectLib.Report(Transition.Area, y)
}
\description{
This function is MPRA_y_SpectLib.Report
}
\examples{
MPRA_y_SpectLib.Report()
}
\keyword{AvantGardeDIA}
