% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{New_Boundaries}
\alias{New_Boundaries}
\title{New_Boundaries}
\usage{
New_Boundaries(L, RowNum_at_max_Skor, Intensity_at_max_Skor)
}
\description{
This function is New_Boundaries
}
\examples{
New_Boundaries()
}
\keyword{AvantGardeDIA}
