% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{num.trans.To.Remove.Fun}
\alias{num.trans.To.Remove.Fun}
\title{num.trans.To.Remove.Fun}
\usage{
num.trans.To.Remove.Fun(num_transX)
}
\description{
This function is num.trans.To.Remove.Fun
}
\examples{
num.trans.To.Remove.Fun()
}
\keyword{AvantGardeDIA}
