% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Rank.filter}
\alias{Rank.filter}
\title{Rank.filter}
\usage{
Rank.filter(y, TopN_RankedbyIntensity, MinimalNumberOfTransitionsAmongTopN,
  Transition.Rank)
}
\description{
This function is Rank.filter
}
\examples{
Rank.filter()
}
\keyword{AvantGardeDIA}
