% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Read_AndFormatResults}
\alias{Read_AndFormatResults}
\title{Read_AndFormatResults}
\usage{
Read_AndFormatResults()
}
\description{
This function reads, concatenates and formats the results.
}
\examples{
Read_AndFormatResults()
}
\keyword{AvantGardeDIA}
