% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Report.Transition.All}
\alias{Report.Transition.All}
\title{Report.Transition.All}
\usage{
Report.Transition.All(Norm.ChromX, y)
}
\description{
This function is Report.Transition.All
}
\examples{
Report.Transition.All()
}
\keyword{AvantGardeDIA}
