% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Report.Transition}
\alias{Report.Transition}
\title{Report.Transition}
\usage{
Report.Transition(Chrom.AnalyteX, Norm.ChromX, y)
}
\description{
This function is Report.Transition
}
\examples{
Report.Transition()
}
\keyword{AvantGardeDIA}
