% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Run_Rescoring_Tool}
\alias{Run_Rescoring_Tool}
\title{Run_Rescoring_Tool}
\usage{
Run_Rescoring_Tool(Chrom.AnalyteX, TransitionRefinementSolution,
  Results_PeakBoundaries_tool, Comment)
}
\description{
This function is Run_Rescoring_Tool
}
\examples{
Run_Rescoring_Tool()
}
\keyword{AvantGardeDIA}
