% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Run_Transition_Refinment_Tool}
\alias{Run_Transition_Refinment_Tool}
\title{Run_Transition_Refinment_Tool}
\usage{
Run_Transition_Refinment_Tool(Chrom.Analyte, Norm.Chrom, MassErrors,
  Transition.Area, Transition.Rank)
}
\description{
This function is Run_Transition_Refinment_Tool
}
\examples{
Run_Transition_Refinment_Tool()
}
\keyword{AvantGardeDIA}
