% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{ScoreTransform.dotP}
\alias{ScoreTransform.dotP}
\title{ScoreTransform.dotP}
\usage{
ScoreTransform.dotP(x)
}
\description{
This function is ScoreTransform.dotP
}
\examples{
ScoreTransform.dotP()
}
\keyword{AvantGardeDIA}
