% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{smooth.Criminal}
\alias{smooth.Criminal}
\title{smooth.Criminal}
\usage{
smooth.Criminal(w)
}
\description{
This function is smooth.Criminal
}
\examples{
smooth.Criminal()
}
\keyword{AvantGardeDIA}
