% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Sug.Matrix.FUN}
\alias{Sug.Matrix.FUN}
\title{Sug.Matrix.FUN}
\usage{
Sug.Matrix.FUN(Trans.Classified, Trans.VectorX)
}
\description{
This function is Sug.Matrix.FUN
}
\examples{
Sug.Matrix.FUN()
}
\keyword{AvantGardeDIA}
