% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Transition.Classifier}
\alias{Transition.Classifier}
\title{Transition.Classifier}
\usage{
Transition.Classifier(Norm.Chrom, Transition.Area, num_trans, Transition.Rank)
}
\description{
This function is Transition.Classifier
}
\examples{
Transition.Classifier()
}
\keyword{AvantGardeDIA}
