% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Transition.Classifier2}
\alias{Transition.Classifier2}
\title{Transition.Classifier2}
\usage{
Transition.Classifier2(Norm.Chrom, Transition.Area, num_trans, y,
  Transition.Rank)
}
\description{
This function is Transition.Classifier2
}
\examples{
Transition.Classifier2()
}
\keyword{AvantGardeDIA}
