% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/20190118_Avant-gardeDIA_v28_CleaningNonUsefulFunctions.R
\name{Transition.Remover}
\alias{Transition.Remover}
\title{Transition.Remover}
\usage{
Transition.Remover(Matrix, y)
}
\description{
This function is Transition.Remover
}
\examples{
Transition.Remover()
}
\keyword{AvantGardeDIA}
